feat: introduce `clientBundle.includeCustomCollections` option, #34
