+++ src/bundle-client.ts
@@ -1,16 +1,34 @@
 import { addTemplate, logger } from '@nuxt/kit'
 import type { IconifyIcon, IconifyJSON } from '@iconify/types'
-import type { ModuleOptions } from './types'
+import type { Nuxt } from 'nuxt/schema'
+import type { CustomCollection, ModuleOptions } from './types'
+import { loadCustomCollection } from './collections'
 
-export function registerClientBundle(options: ModuleOptions) {
+export function registerClientBundle(
+  options: ModuleOptions,
+  nuxt: Nuxt,
+) {
   const iconifyCollectionMap = new Map<string, Promise<IconifyJSON | undefined>>()
+
+  const {
+    includeCustomCollections = options.provider !== 'server',
+  } = options.clientBundle || {}
 
   // Client bundle
   addTemplate({
     filename: 'nuxt-icon-client-bundle.mjs',
     write: true,
     async getContents() {
-      const icons = options.clientBundle?.icons || []
+      const icons = [...options.clientBundle?.icons || []]
+
+      let customCollections: IconifyJSON[] = []
+      if (includeCustomCollections && options.customCollections?.length) {
+        customCollections = await Promise.all(
+          options.customCollections.map(collection => loadCustomCollection(collection, nuxt)),
+        )
+      }
+
+      console.log({ icons })
 
       if (!icons.length)
         return 'export function init() {}'
@@ -41,6 +59,10 @@ export function registerClientBundle(options: ModuleOptions) {
             return `  /* ${icon} failed to load */`
           }
           return `  addIcon('${icon}', ${JSON.stringify(data)})`
+        })),
+        customCollections.length ? '  // ===== Custom collections =====' : '',
+        ...customCollections.flatMap(collection => Object.entries(collection.icons).map(([name, data]) => {
+          return `  addIcon('${collection.prefix}:${name}', ${JSON.stringify(data)})`
         })),
         '  _initialized = true',
         '}',
