+++ src/collection-names.ts
@@ -1,0 +1,179 @@
+// GENERATED BY scripts/collections.ts
+export default [
+  "academicons",
+  "akar-icons",
+  "ant-design",
+  "arcticons",
+  "basil",
+  "bi",
+  "bitcoin-icons",
+  "bpmn",
+  "brandico",
+  "bx",
+  "bxl",
+  "bxs",
+  "bytesize",
+  "carbon",
+  "catppuccin",
+  "cbi",
+  "charm",
+  "ci",
+  "cib",
+  "cif",
+  "cil",
+  "circle-flags",
+  "circum",
+  "clarity",
+  "codicon",
+  "covid",
+  "cryptocurrency",
+  "cryptocurrency-color",
+  "dashicons",
+  "devicon",
+  "devicon-plain",
+  "ei",
+  "el",
+  "emojione",
+  "emojione-monotone",
+  "emojione-v1",
+  "entypo",
+  "entypo-social",
+  "eos-icons",
+  "ep",
+  "et",
+  "eva",
+  "f7",
+  "fa",
+  "fa-brands",
+  "fa-regular",
+  "fa-solid",
+  "fa6-brands",
+  "fa6-regular",
+  "fa6-solid",
+  "fad",
+  "fe",
+  "feather",
+  "file-icons",
+  "flag",
+  "flagpack",
+  "flat-color-icons",
+  "flat-ui",
+  "flowbite",
+  "fluent",
+  "fluent-emoji",
+  "fluent-emoji-flat",
+  "fluent-emoji-high-contrast",
+  "fluent-mdl2",
+  "fontelico",
+  "fontisto",
+  "formkit",
+  "foundation",
+  "fxemoji",
+  "gala",
+  "game-icons",
+  "geo",
+  "gg",
+  "gis",
+  "gravity-ui",
+  "gridicons",
+  "grommet-icons",
+  "guidance",
+  "healthicons",
+  "heroicons",
+  "heroicons-outline",
+  "heroicons-solid",
+  "hugeicons",
+  "humbleicons",
+  "ic",
+  "icomoon-free",
+  "icon-park",
+  "icon-park-outline",
+  "icon-park-solid",
+  "icon-park-twotone",
+  "iconamoon",
+  "iconoir",
+  "icons8",
+  "il",
+  "ion",
+  "iwwa",
+  "jam",
+  "la",
+  "lets-icons",
+  "line-md",
+  "logos",
+  "ls",
+  "lucide",
+  "mage",
+  "majesticons",
+  "maki",
+  "map",
+  "marketeq",
+  "material-symbols",
+  "material-symbols-light",
+  "mdi",
+  "mdi-light",
+  "medical-icon",
+  "memory",
+  "meteocons",
+  "mi",
+  "mingcute",
+  "mono-icons",
+  "mynaui",
+  "nimbus",
+  "nonicons",
+  "noto",
+  "noto-v1",
+  "octicon",
+  "oi",
+  "ooui",
+  "openmoji",
+  "oui",
+  "pajamas",
+  "pepicons",
+  "pepicons-pencil",
+  "pepicons-pop",
+  "pepicons-print",
+  "ph",
+  "pixelarticons",
+  "prime",
+  "ps",
+  "quill",
+  "radix-icons",
+  "raphael",
+  "ri",
+  "rivet-icons",
+  "si-glyph",
+  "simple-icons",
+  "simple-line-icons",
+  "skill-icons",
+  "solar",
+  "streamline",
+  "streamline-emojis",
+  "subway",
+  "svg-spinners",
+  "system-uicons",
+  "tabler",
+  "tdesign",
+  "teenyicons",
+  "token",
+  "token-branded",
+  "topcoat",
+  "twemoji",
+  "typcn",
+  "uil",
+  "uim",
+  "uis",
+  "uit",
+  "uiw",
+  "unjs",
+  "vaadin",
+  "vs",
+  "vscode-icons",
+  "websymbol",
+  "weui",
+  "whh",
+  "wi",
+  "wpf",
+  "zmdi",
+  "zondicons"
+]
