+++ src/collections.ts
@@ -1,175 +1,70 @@
-// GENERATED BY scripts/collections.ts
-export default [
-  "academicons",
-  "akar-icons",
-  "ant-design",
-  "arcticons",
-  "basil",
-  "bi",
-  "bitcoin-icons",
-  "bpmn",
-  "brandico",
-  "bx",
-  "bxl",
-  "bxs",
-  "bytesize",
-  "carbon",
-  "cbi",
-  "charm",
-  "ci",
-  "cib",
-  "cif",
-  "cil",
-  "circle-flags",
-  "circum",
-  "clarity",
-  "codicon",
-  "covid",
-  "cryptocurrency",
-  "cryptocurrency-color",
-  "dashicons",
-  "devicon",
-  "devicon-plain",
-  "ei",
-  "el",
-  "emojione",
-  "emojione-monotone",
-  "emojione-v1",
-  "entypo",
-  "entypo-social",
-  "eos-icons",
-  "ep",
-  "et",
-  "eva",
-  "f7",
-  "fa",
-  "fa-brands",
-  "fa-regular",
-  "fa-solid",
-  "fa6-brands",
-  "fa6-regular",
-  "fa6-solid",
-  "fad",
-  "fe",
-  "feather",
-  "file-icons",
-  "flag",
-  "flagpack",
-  "flat-color-icons",
-  "flat-ui",
-  "flowbite",
-  "fluent",
-  "fluent-emoji",
-  "fluent-emoji-flat",
-  "fluent-emoji-high-contrast",
-  "fluent-mdl2",
-  "fontelico",
-  "fontisto",
-  "formkit",
-  "foundation",
-  "fxemoji",
-  "gala",
-  "game-icons",
-  "geo",
-  "gg",
-  "gis",
-  "gravity-ui",
-  "gridicons",
-  "grommet-icons",
-  "guidance",
-  "healthicons",
-  "heroicons",
-  "heroicons-outline",
-  "heroicons-solid",
-  "humbleicons",
-  "ic",
-  "icomoon-free",
-  "icon-park",
-  "icon-park-outline",
-  "icon-park-solid",
-  "icon-park-twotone",
-  "iconamoon",
-  "iconoir",
-  "icons8",
-  "il",
-  "ion",
-  "iwwa",
-  "jam",
-  "la",
-  "lets-icons",
-  "line-md",
-  "logos",
-  "ls",
-  "lucide",
-  "mage",
-  "majesticons",
-  "maki",
-  "map",
-  "marketeq",
-  "material-symbols",
-  "material-symbols-light",
-  "mdi",
-  "mdi-light",
-  "medical-icon",
-  "memory",
-  "meteocons",
-  "mi",
-  "mingcute",
-  "mono-icons",
-  "mynaui",
-  "nimbus",
-  "nonicons",
-  "noto",
-  "noto-v1",
-  "octicon",
-  "oi",
-  "ooui",
-  "openmoji",
-  "oui",
-  "pajamas",
-  "pepicons",
-  "pepicons-pencil",
-  "pepicons-pop",
-  "pepicons-print",
-  "ph",
-  "pixelarticons",
-  "prime",
-  "ps",
-  "quill",
-  "radix-icons",
-  "raphael",
-  "ri",
-  "si-glyph",
-  "simple-icons",
-  "simple-line-icons",
-  "skill-icons",
-  "solar",
-  "streamline",
-  "streamline-emojis",
-  "subway",
-  "svg-spinners",
-  "system-uicons",
-  "tabler",
-  "tdesign",
-  "teenyicons",
-  "token",
-  "token-branded",
-  "topcoat",
-  "twemoji",
-  "typcn",
-  "uil",
-  "uim",
-  "uis",
-  "uit",
-  "uiw",
-  "unjs",
-  "vaadin",
-  "vs",
-  "vscode-icons",
-  "websymbol",
-  "whh",
-  "wi",
-  "wpf",
-  "zmdi",
-  "zondicons"
-]
+import { basename, join, isAbsolute } from 'node:path'
+import fs from 'node:fs/promises'
+import { logger } from '@nuxt/kit'
+import type { Nuxt } from '@nuxt/schema'
+import fg from 'fast-glob'
+import type { IconifyJSON } from '@iconify/types'
+import { parseSVGContent, convertParsedSVG } from '@iconify/utils/lib/svg/parse'
+import { isPackageExists } from 'local-pkg'
+import collectionNames from './collection-names'
+import type { CustomCollection, ServerBundleOptions, RemoteCollection } from './types'
+
+export const isFullCollectionExists = isPackageExists('@iconify/json')
+
+export async function resolveCollection(
+  nuxt: Nuxt,
+  collection: string | IconifyJSON | CustomCollection | RemoteCollection,
+): Promise<string | IconifyJSON | RemoteCollection> {
+  if (typeof collection === 'string')
+    return collection
+  // Custom collection
+  if ('dir' in collection) {
+    return await loadCustomCollection(collection, nuxt)
+  }
+  return collection
+}
+
+export async function loadCustomCollection(collection: CustomCollection, nuxt: Nuxt): Promise<IconifyJSON> {
+  const dir = isAbsolute(collection.dir) ? collection.dir : join(nuxt.options.rootDir, collection.dir)
+  const files = (await fg('*.svg', { cwd: dir, onlyFiles: true }))
+    .sort()
+
+  const parsedIcons = await Promise.all(files.map(async (file) => {
+    const name = basename(file, '.svg')
+    let svg = await fs.readFile(join(dir, file), 'utf-8')
+    const cleanupIdx = svg.indexOf('<svg')
+    if (cleanupIdx > 0)
+      svg = svg.slice(cleanupIdx)
+    const data = convertParsedSVG(parseSVGContent(svg)!)
+    if (!data) {
+      logger.error(`Nuxt Icon could not parse the SVG content for icon \`${name}\``)
+      return [name, {}]
+    }
+    if (data.top === 0)
+      delete data.top
+    if (data.left === 0)
+      delete data.left
+    return [name, data]
+  }))
+
+  const successfulIcons = parsedIcons.filter(([_, data]) => Object.keys(data).length > 0)
+  // @ts-expect-error remove extra properties
+  delete collection.dir
+
+  logger.success(`Nuxt Icon loaded local collection \`${collection.prefix}\` with ${successfulIcons.length} icons`)
+  return {
+    ...collection,
+    icons: Object.fromEntries(successfulIcons),
+  }
+}
+
+export async function discoverInstalledCollections(): Promise<ServerBundleOptions['collections']> {
+  const collections = isFullCollectionExists
+    ? collectionNames
+    : collectionNames.filter(collection => isPackageExists('@iconify-json/' + collection))
+  if (isFullCollectionExists)
+    logger.success(`Nuxt Icon discovered local-installed ${collections.length} collections (@iconify/json)`)
+  else if (collections.length)
+    logger.success(`Nuxt Icon discovered local-installed ${collections.length} collections:`, collections.join(', '))
+  return collections
+}
